+++ scripts/build_industry_sector_ratios.py
@@ -60,6 +60,7 @@ index = [
     "hydrogen",
     "heat",
     "naphtha",
+    "ammonia",
     "process emission",
     "process emission from feedstock",
 ]
@@ -432,8 +433,11 @@ def chemicals_industry():
 
     sector = "Ammonia"
     df[sector] = 0.0
-    df.loc["hydrogen", sector] = config["MWh_H2_per_tNH3_electrolysis"]
-    df.loc["elec", sector] = config["MWh_elec_per_tNH3_electrolysis"]
+    if snakemake.config["sector"].get("ammonia", False):
+        df.loc["ammonia", sector] = config["MWh_NH3_per_tNH3"]
+    else:
+        df.loc["hydrogen", sector] = config["MWh_H2_per_tNH3_electrolysis"]
+        df.loc["elec", sector] = config["MWh_elec_per_tNH3_electrolysis"]
 
     # Chlorine
 
@@ -614,7 +618,7 @@ def nonmetalic_mineral_products():
     # (c) clinker production (kilns),
     # (d) Grinding, packaging.
     # (b)+(c) represent 94% of fec. So (a) is joined to (b) and (d) is joined to (c).
-    # Temperatures above 1400C are required for procesing limestone and sand into clinker.
+    # Temperatures above 1400C are required for processing limestone and sand into clinker.
     # Everything (except current electricity and heat consumption and existing biomass)
     # is transformed into methane for high T.
 
@@ -1106,7 +1110,7 @@ def non_ferrous_metals():
 
     # Aluminium secondary route
 
-    # All is coverted into secondary route fully electrified.
+    # All is converted into secondary route fully electrified.
 
     sector = "Aluminium - secondary production"