+++ scripts/helper.py
@@ -25,7 +25,7 @@ def override_component_attrs(directory):
 
     Returns
     -------
-    Dictionary of overriden component attributes.
+    Dictionary of overridden component attributes.
     """
 
     attrs = Dict({k : v.copy() for k,v in component_attrs.items()})