+++ scripts/make_summary.py
@@ -273,7 +273,7 @@ def calculate_supply(n, label, supply):
 
             for end in [col[3:] for col in c.df.columns if col[:3] == "bus"]:
 
-                items = c.df.index[c.df["bus" + end].map(bus_map, na_action=False)]
+                items = c.df.index[c.df["bus" + end].map(bus_map, na_action=None)]
 
                 if len(items) == 0:
                     continue
@@ -318,7 +318,7 @@ def calculate_supply_energy(n, label, supply_energy):
 
             for end in [col[3:] for col in c.df.columns if col[:3] == "bus"]:
 
-                items = c.df.index[c.df["bus" + str(end)].map(bus_map, na_action=False)]
+                items = c.df.index[c.df["bus" + str(end)].map(bus_map, na_action=None)]
 
                 if len(items) == 0:
                     continue